Yet another refactoring
